♻ refactored common models/utils
